Cast ptrdiff_t value to int for %.*s argument.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@858 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/io_lib/utils/hash_table.c
@@ -971,7 +971,7 @@ HashFile *HashFileOpen(char *fname) {
 		HashFileDestroy(hf);
 		return NULL;
 	    }
-	    sprintf(aname, "%.*s%s", cp-fname+1, fname, hf->archive);
+	    sprintf(aname, "%.*s%s", (int)(cp-fname+1), fname, hf->archive);
 	    if (NULL == (hf->afp = fopen(aname, "rb"))) {
 
 		return NULL;
